Move useful integration utils to main library
Moving the Agent and Client pytest fixtures to the main library will allow use
in external Agent development, such as socs.
