+++ tests/integration/test_aggregator_agent_integration.py
@@ -3,9 +3,12 @@ import pytest
 import ocs
 from ocs.base import OpCode
 
-from integration.util import (
+from ocs.testing import (
     create_agent_runner_fixture,
     create_client_fixture,
+)
+
+from integration.util import (
     create_crossbar_fixture
 )
