+++ tests/integration/test_fake_data_agent_integration.py
@@ -3,10 +3,13 @@ import pytest
 import ocs
 from ocs.base import OpCode
 
-from integration.util import (
+from ocs.testing import (
     create_agent_runner_fixture,
     create_client_fixture,
-    create_crossbar_fixture
+)
+
+from integration.util import (
+    create_crossbar_fixture,
 )
 
 pytest_plugins = ("docker_compose")
